initialise tables

+++ src/main/java/gr/aueb/cf/eshopfinalproject/repository/SalesRepository.java
@@ -3,5 +3,5 @@ package gr.aueb.cf.eshopfinalproject.repository;
 import gr.aueb.cf.eshopfinalproject.model.Sales;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalesRepository extends JpaRepository<Sales, Integer> {
+public interface SalesRepository extends JpaRepository<Sales, Long> {
 }
